refactor(balances): limit_reached cap selection in its own files
