+++ server/src/internal/balances/trackWebhooks/limitReached/types/blockingUsageLimit.ts
@@ -1,0 +1,6 @@
+import type { UsageLimitFilter, UsageLimitWebhookBlock } from "@autumn/shared";
+
+export type BlockingUsageLimit = {
+	block: UsageLimitWebhookBlock;
+	filter: UsageLimitFilter | undefined;
+};
